# docs/swagger.yaml
basePath: /
definitions:
  serializer.AddToCalculationRequest:
    properties:
      bread_units:
        type: number
      current_glucose:
        type: number
    required:
    - bread_units
    - current_glucose
    type: object
...
  serializer.PatientJSON:
    properties:
      description:
        type: string
      glucose:
        type: number
      name:
        type: string
      patient_id:
        type: integer
      photo_url:
        type: string
      sensitivity:
        type: number
      status:
        type: string
      type:
        type: integer
    required:
    - glucose
    - name
    - sensitivity
    - type
    type: object
  serializer.PatientRespJSON:
    properties:
      description:
        type: string
      glucose:
        type: number
      name:
        type: string
      patient_id:
        type: integer
      photo_url:
        type: string
      sensitivity:
        type: number
      status:
        type: string
      type:
        type: integer
    type: object
...
host: localhost:8080
info:
  contact:
    email: support@insulin.ru
    name: API Support
  description: API для расчета болюсного инсулина
  license:
    name: AS IS (NO WARRANTY)
  title: Insulin Calculation API
  version: "1.0"
paths:
...
  /api/patients:
    get:
      consumes:
      - application/json
      description: Возвращает список пациентов с возможностью фильтрации
      parameters:
      - description: Фильтр по имени
        in: query
        name: name
        type: string
      - description: Фильтр по типу пациента (1,2,3)
        in: query
        name: type
        type: string
      - description: Фильтр по статусу
        in: query
        name: status
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Список пациентов
          schema:
            items:
              $ref: '#/definitions/serializer.PatientRespJSON'
            type: array
        "500":
          description: Внутренняя ошибка сервера
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - ApiKeyAuth: []
      summary: Получить список пациентов
      tags:
      - patients
    post:
      consumes:
      - application/json
      description: Создает нового пациента в системе
      parameters:
      - description: Данные пациента
        in: body
        name: patient
        required: true
        schema:
          $ref: '#/definitions/serializer.PatientJSON'
      produces:
      - application/json
      responses:
        "201":
          description: Пациент успешно создан
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Ошибка валидации
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Внутренняя ошибка сервера
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - ApiKeyAuth: []
      summary: Создать нового пациента
      tags:
      - patients
  /api/patients/{patient_id}:
    delete:
      consumes:
      - application/json
      description: Помечает пациента как удаленного (мягкое удаление)
      parameters:
      - description: ID пациента
        in: path
        name: patient_id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Пациент успешно удален
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Неверный ID пациента
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Пациент не найден
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Внутренняя ошибка сервера
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - ApiKeyAuth: []
      summary: Удалить пациента
      tags:
      - patients
    get:
      consumes:
      - application/json
      description: Возвращает данные конкретного пациента
      parameters:
      - description: ID пациента
        in: path
        name: patient_id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Данные пациента
          schema:
            $ref: '#/definitions/serializer.PatientRespJSON'
        "400":
          description: Неверный ID пациента
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Пациент не найден
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - ApiKeyAuth: []
      summary: Получить пациента по ID
      tags:
      - patients
    put:
      consumes:
      - application/json
      description: Обновляет данные существующего пациента
      parameters:
      - description: ID пациента
        in: path
        name: patient_id
        required: true
        type: integer
      - description: Новые данные пациента
        in: body
        name: patient
        required: true
        schema:
          $ref: '#/definitions/serializer.PatientJSON'
      produces:
      - application/json
      responses:
        "200":
          description: Пациент успешно обновлен
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Ошибка валидации
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Пациент не найден
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Внутренняя ошибка сервера
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - ApiKeyAuth: []
      summary: Обновить данные пациента
      tags:
      - patients
  /api/patients/{patient_id}/add:
    post:
      consumes:
      - application/json
      description: Добавляет пациента в заявку-черновик для расчета инсулина
      parameters:
      - description: ID пациента
        in: path
        name: patient_id
        required: true
        type: integer
      - description: Данные для расчета
        in: body
        name: calculation
        required: true
        schema:
          $ref: '#/definitions/serializer.AddToCalculationRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Пациент успешно добавлен в расчет
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Ошибка валидации
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Пациент не найден
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Внутренняя ошибка сервера
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - ApiKeyAuth: []
      summary: Добавить пациента в расчет инсулина
      tags:
      - patients
  /api/patients/{patient_id}/photo:
    post:
      consumes:
      - multipart/form-data
      description: Загружает фотографию для пациента
      parameters:
      - description: ID пациента
        in: path
        name: patient_id
        required: true
        type: integer
      - description: Фото пациента
        in: formData
        name: photo
        required: true
        type: file
      produces:
      - application/json
      responses:
        "200":
          description: Фото успешно загружено
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Ошибка валидации
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Пациент не найден
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Внутренняя ошибка сервера
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - ApiKeyAuth: []
      summary: Загрузить фото пациента
      tags:
      - patients
...
schemes:
- https
- http
securityDefinitions:
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"